Add files via upload
This is a Lorem Ipsum generator. It is supposed to make random sentences and paragraphs from an array of words (mine happens to be pizza themed). I used a method to create a sentence, and then a method to make paragraphs (with a random amount of sentences) and then the user gets to decide how many paragraphs of text they would like (which is executed with a method). The final will have more words to choose from, this is just a test version.